LOCAL: Disable the xterm terminal emulation
It does the correct thing if there is already a soc_term running.
However, because I don't run on my local machine, this results either
in windows in weird places, or windows over a slow, forwarded X11
connection.

I'll leave it to myself to start the soc_term processes after sshing
into the development machine.